+++ src/Stormancer.Plugins/Epic/Stormancer.Server.Plugins.Epic/Dto/Account.cs
@@ -1,7 +1,7 @@
 using System.Collections.Generic;
 using System.Text.Json.Serialization;
 
-namespace Stormancer.Server.Plugins.Epic
+namespace Stormancer.Server.Plugins.Eos
 {
     /// <summary>
     /// Linked account