+++ src/Stormancer.Plugins/Epic/Stormancer.Server.Plugins.Epic/EosFriendsEventHandler.cs
@@ -6,18 +6,18 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Stormancer.Server.Plugins.Epic
+namespace Stormancer.Server.Plugins.Eos
 {
-    internal class EpicFriendsEventHandler : IFriendsEventHandler
+    internal class EosFriendsEventHandler : IFriendsEventHandler
     {
 
         Task IFriendsEventHandler.OnAddingFriend(AddingFriendCtx ctx)
         {
             foreach (var friend in ctx.Friends)
             {
-                if (friend.userInfos?.User != null && !friend.friend.UserIds.Any(p => p.Platform == EpicConstants.PLATFORM_NAME) && friend.userInfos.User.TryGetEpicAccountId(out var accountId))
+                if (friend.userInfos?.User != null && !friend.friend.UserIds.Any(p => p.Platform == Eos.PLATFORM_NAME) && friend.userInfos.User.TryGetEpicAccountId(out var accountId))
                 {
-                    friend.friend.UserIds.Add(new PlatformId(EpicConstants.PLATFORM_NAME, accountId));
+                    friend.friend.UserIds.Add(new PlatformId(Eos.PLATFORM_NAME, accountId));
                 }
             }